requirements.txt
setuptools>=65.5.1
google-cloud-translate==3.12.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
pdfplumber==0.10.3
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pandas==2.1.4
tqdm==4.66.1 